pyproject.toml
[tool.poetry]
name = "connectnow"
version = "0.1.0"
description = ""
authors = ["thenesterov <thenesterov@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
SQLAlchemy = "^1.4.42"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"
pydantic = {extras = ["email"], version = "^1.10.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
